Split room manager and allow overrides
