+++ osu.Game.Tests/Visual/Multiplayer/TestSceneLoungeRoomsContainer.cs
@@ -9,6 +9,7 @@ using osu.Game.Graphics;
 using osu.Game.Online.Rooms;
 using osu.Game.Rulesets.Catch;
 using osu.Game.Rulesets.Osu;
+using osu.Game.Screens.OnlinePlay;
 using osu.Game.Screens.OnlinePlay.Lounge.Components;
 using osu.Game.Tests.Visual.OnlinePlay;
 using osuTK.Graphics;
@@ -18,6 +19,8 @@ namespace osu.Game.Tests.Visual.Multiplayer
 {
     public class TestSceneLoungeRoomsContainer : OnlinePlaySubScreenTestScene
     {
+        protected new TestBasicRoomManager RoomManager => (TestBasicRoomManager)base.RoomManager;
+
         private RoomsContainer container;
 
         [BackgroundDependencyLoader]
@@ -31,6 +34,8 @@ namespace osu.Game.Tests.Visual.Multiplayer
                 JoinRequested = joinRequested
             };
         }
+
+        protected override IRoomManager CreateRoomManager() => new TestBasicRoomManager();
 
         [Test]
         public void TestBasicListChanges()
